+++ integration/k3t/k3t.go
@@ -1,0 +1,2 @@
+// Package k3t provides a simple wrapper around the k3d client to operate a cluster for integration tests.
+package k3t