+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## main / unreleased
 
 * [ENHANCEMENT] Update Go to `1.19.4`. #29
+* [ENHANCEMENT] Added `-reconcile.interval` option to configure the minimum reconciliation interval. #30  
 
 ## v0.2.0